imx6: debian: update emmc uboot instructions with boot partition select
Add to debian bootloader installation an extra step for selecting the
eMMC boot partition. Factory default is numeric 0 meaning disabled, but
it happened to work with data partition.
Newer units may come with a default of 1 meaning boot0 partition.
Therefore add the extra step to ensure changing to 7 for the data
partition.